requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
pandas==2.2.2
openpyxl==3.1.5
pytz==2024.1
python-dotenv==1.0.1
